Moved each pass into its own module
